Refactor URL helpers to injectable service and upgrade packages to get access to 2.0.1 of AutoConfiguration
