+++ src/SFA.DAS.PAS.Account.Api.UnitTests/packages.config
@@ -4,7 +4,7 @@
   <package id="Castle.Core" version="3.3.3" targetFramework="net45" />
   <package id="Fare" version="2.1.1" targetFramework="net462" />
   <package id="FluentAssertions" version="4.15.0" targetFramework="net45" />
-  <package id="MediatR" version="5.1.0" targetFramework="net462" />
+  <package id="MediatR" version="6.0.0" targetFramework="net462" />
   <package id="Moq" version="4.5.28" targetFramework="net45" />
   <package id="NUnit" version="3.10.1" targetFramework="net462" />
   <package id="NUnit3TestAdapter" version="3.10.0" targetFramework="net462" />
