+++ src/SFA.DAS.PAS.Account.Api/packages.config
@@ -4,7 +4,7 @@
   <package id="AutoMapper" version="6.1.1" targetFramework="net45" />
   <package id="CommonServiceLocator" version="1.3" targetFramework="net45" />
   <package id="FluentValidation" version="8.0.0" targetFramework="net462" />
-  <package id="MediatR" version="5.1.0" targetFramework="net462" />
+  <package id="MediatR" version="6.0.0" targetFramework="net462" />
   <package id="Microsoft.ApplicationInsights" version="2.5.1" targetFramework="net462" />
   <package id="Microsoft.ApplicationInsights.Agent.Intercept" version="2.4.0" targetFramework="net45" />
   <package id="Microsoft.ApplicationInsights.DependencyCollector" version="2.5.1" targetFramework="net462" />
