+++ src/SFA.DAS.ProviderApprenticeshipsService.Application.UnitTests/packages.config
@@ -9,11 +9,11 @@
   <package id="FluentAssertions" version="4.15.0" targetFramework="net452" />
   <package id="FluentValidation" version="8.0.0" targetFramework="net462" />
   <package id="Hashids.net" version="1.2.2" targetFramework="net462" />
-  <package id="MediatR" version="5.1.0" targetFramework="net462" />
+  <package id="MediatR" version="6.0.0" targetFramework="net462" />
   <package id="Microsoft.AspNet.WebApi.Client" version="5.2.6" targetFramework="net462" />
   <package id="Microsoft.AspNet.WebApi.Core" version="5.2.6" targetFramework="net462" />
   <package id="Microsoft.AspNetCore.WebUtilities" version="2.1.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.DocumentDB.Core" version="2.1.3" targetFramework="net462" />
+  <package id="Microsoft.Azure.DocumentDB.Core" version="2.2.1" targetFramework="net462" />
   <package id="Microsoft.Azure.KeyVault.Core" version="1.0.0" targetFramework="net462" />
   <package id="Microsoft.Extensions.Primitives" version="2.1.1" targetFramework="net462" />
   <package id="Microsoft.IdentityModel.Clients.ActiveDirectory" version="3.19.8" targetFramework="net462" />
@@ -29,25 +29,26 @@
   <package id="NUnit3TestAdapter" version="3.10.0" targetFramework="net462" />
   <package id="SFA.DAS.Apprenticeships.Api.Client" version="0.11.75" targetFramework="net462" />
   <package id="SFA.DAS.Apprenticeships.Api.Types" version="0.11.82" targetFramework="net462" />
-  <package id="SFA.DAS.AutoConfiguration" version="1.1.1" targetFramework="net462" />
+  <package id="SFA.DAS.AutoConfiguration" version="2.0.1" targetFramework="net462" />
   <package id="SFA.DAS.Commitments.Api.Client" version="2.1.716" targetFramework="net462" />
   <package id="SFA.DAS.Commitments.Api.Types" version="2.1.716" targetFramework="net462" />
   <package id="SFA.DAS.Common.Domain" version="1.4.43" targetFramework="net462" />
-  <package id="SFA.DAS.CosmosDb" version="1.0.1" targetFramework="net462" />
+  <package id="SFA.DAS.CosmosDb" version="1.0.47" targetFramework="net462" />
   <package id="SFA.DAS.CosmosDb.Testing" version="1.0.1" targetFramework="net462" />
   <package id="SFA.DAS.HashingService" version="1.0.0.42526" targetFramework="net462" />
   <package id="SFA.DAS.Http" version="1.2.4" targetFramework="net462" />
   <package id="SFA.DAS.NLog.Logger" version="1.0.0.39208" targetFramework="net462" />
   <package id="SFA.DAS.Notifications.Api.Client" version="2.0.0.35919" targetFramework="net452" />
   <package id="SFA.DAS.Notifications.Api.Types" version="2.0.0.35919" targetFramework="net452" />
-  <package id="SFA.DAS.ProviderRelationships.Api.Client" version="1.1.1" targetFramework="net462" />
+  <package id="SFA.DAS.ProviderRelationships.Api.Client" version="2.0.1" targetFramework="net462" />
   <package id="SFA.DAS.ProviderRelationships.ReadStore" version="1.1.1" targetFramework="net462" />
-  <package id="SFA.DAS.ProviderRelationships.Types" version="1.1.1" targetFramework="net462" />
+  <package id="SFA.DAS.ProviderRelationships.Types" version="2.0.1" targetFramework="net462" />
   <package id="StructureMap" version="4.7.0" targetFramework="net462" />
   <package id="System.AppContext" version="4.1.0" targetFramework="net462" />
   <package id="System.Buffers" version="4.5.0" targetFramework="net462" />
   <package id="System.Collections" version="4.0.11" targetFramework="net462" />
   <package id="System.Collections.Concurrent" version="4.3.0" targetFramework="net462" />
+  <package id="System.Collections.Immutable" version="1.3.0" targetFramework="net462" />
   <package id="System.Collections.NonGeneric" version="4.0.1" targetFramework="net462" />
   <package id="System.Collections.Specialized" version="4.0.1" targetFramework="net462" />
   <package id="System.ComponentModel.Annotations" version="4.4.1" targetFramework="net462" />
@@ -69,7 +70,7 @@
   <package id="System.Linq.Expressions" version="4.3.0" targetFramework="net462" />
   <package id="System.Linq.Queryable" version="4.0.1" targetFramework="net462" />
   <package id="System.Memory" version="4.5.1" targetFramework="net462" />
-  <package id="System.Net.Http" version="4.3.3" targetFramework="net462" />
+  <package id="System.Net.Http" version="4.3.4" targetFramework="net462" />
   <package id="System.Net.NameResolution" version="4.0.0" targetFramework="net462" />
   <package id="System.Net.NetworkInformation" version="4.1.0" targetFramework="net462" />
   <package id="System.Net.Primitives" version="4.0.11" targetFramework="net462" />
