+++ src/SFA.DAS.ProviderApprenticeshipsService.Infrastructure/packages.config
@@ -5,10 +5,11 @@
   <package id="FeatureToggle" version="4.0.2" targetFramework="net462" />
   <package id="FeatureToggle.Common" version="4.0.2" targetFramework="net462" />
   <package id="Hashids.net" version="1.2.2" targetFramework="net45" />
+  <package id="MediatR" version="6.0.0" targetFramework="net462" />
   <package id="Microsoft.AspNet.WebApi.Client" version="5.2.6" targetFramework="net462" />
   <package id="Microsoft.AspNet.WebApi.Core" version="5.2.6" targetFramework="net462" />
   <package id="Microsoft.AspNetCore.WebUtilities" version="2.1.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.DocumentDB.Core" version="2.1.3" targetFramework="net462" />
+  <package id="Microsoft.Azure.DocumentDB.Core" version="2.2.1" targetFramework="net462" />
   <package id="Microsoft.Azure.KeyVault.Core" version="1.0.0" targetFramework="net45" />
   <package id="Microsoft.Data.Edm" version="5.8.0" targetFramework="net45" />
   <package id="Microsoft.Data.OData" version="5.8.0" targetFramework="net45" />
@@ -25,7 +26,7 @@
   <package id="Polly" version="5.1.0" targetFramework="net45" />
   <package id="SFA.DAS.Account.Api.Client" version="1.3.928" targetFramework="net462" />
   <package id="SFA.DAS.Account.Api.Types" version="1.3.928" targetFramework="net462" />
-  <package id="SFA.DAS.AutoConfiguration" version="1.1.1" targetFramework="net462" />
+  <package id="SFA.DAS.AutoConfiguration" version="2.0.1" targetFramework="net462" />
   <package id="SFA.DAS.Commitments.Api.Client" version="2.1.716" targetFramework="net462" />
   <package id="SFA.DAS.Commitments.Api.Types" version="2.1.716" targetFramework="net462" />
   <package id="SFA.DAS.Common.Domain" version="1.4.43" targetFramework="net462" />
@@ -33,16 +34,16 @@
   <package id="SFA.DAS.Configuration.AzureTableStorage" version="1.0.0.53229" targetFramework="net462" />
   <package id="SFA.DAS.Configuration.FileStorage" version="1.0.0.53229" targetFramework="net462" />
   <package id="SFA.DAS.CookieService" version="1.0.0.27423" targetFramework="net45" />
-  <package id="SFA.DAS.CosmosDb" version="1.0.1" targetFramework="net462" />
+  <package id="SFA.DAS.CosmosDb" version="1.0.47" targetFramework="net462" />
   <package id="SFA.DAS.CosmosDb.Testing" version="1.0.1" targetFramework="net462" />
   <package id="SFA.DAS.HashingService" version="1.0.0.42526" targetFramework="net462" />
   <package id="SFA.DAS.Http" version="1.2.4" targetFramework="net462" />
   <package id="SFA.DAS.NLog.Logger" version="1.0.0.39208" targetFramework="net462" />
   <package id="SFA.DAS.Notifications.Api.Client" version="2.0.0.35919" targetFramework="net45" />
   <package id="SFA.DAS.Notifications.Api.Types" version="2.0.0.35919" targetFramework="net45" />
-  <package id="SFA.DAS.ProviderRelationships.Api.Client" version="1.1.1" targetFramework="net462" />
+  <package id="SFA.DAS.ProviderRelationships.Api.Client" version="2.0.1" targetFramework="net462" />
   <package id="SFA.DAS.ProviderRelationships.ReadStore" version="1.1.1" targetFramework="net462" />
-  <package id="SFA.DAS.ProviderRelationships.Types" version="1.1.1" targetFramework="net462" />
+  <package id="SFA.DAS.ProviderRelationships.Types" version="2.0.1" targetFramework="net462" />
   <package id="SFA.DAS.Sql.Client" version="1.0.0.32930" targetFramework="net45" />
   <package id="StackExchange.Redis" version="1.2.6" targetFramework="net462" />
   <package id="StructureMap" version="4.7.0" targetFramework="net462" />
@@ -50,6 +51,7 @@
   <package id="System.Buffers" version="4.5.0" targetFramework="net462" />
   <package id="System.Collections" version="4.3.0" targetFramework="net462" />
   <package id="System.Collections.Concurrent" version="4.3.0" targetFramework="net462" />
+  <package id="System.Collections.Immutable" version="1.3.0" targetFramework="net462" />
   <package id="System.Collections.NonGeneric" version="4.0.1" targetFramework="net462" />
   <package id="System.Collections.Specialized" version="4.0.1" targetFramework="net462" />
   <package id="System.Console" version="4.3.0" targetFramework="net462" />
@@ -70,7 +72,7 @@
   <package id="System.Linq.Expressions" version="4.3.0" targetFramework="net462" />
   <package id="System.Linq.Queryable" version="4.0.1" targetFramework="net462" />
   <package id="System.Memory" version="4.5.1" targetFramework="net462" />
-  <package id="System.Net.Http" version="4.3.3" targetFramework="net462" />
+  <package id="System.Net.Http" version="4.3.4" targetFramework="net462" />
   <package id="System.Net.NameResolution" version="4.0.0" targetFramework="net462" />
   <package id="System.Net.NetworkInformation" version="4.1.0" targetFramework="net462" />
   <package id="System.Net.Primitives" version="4.3.0" targetFramework="net462" />
