+++ src/SFA.DAS.ProviderApprenticeshipsService.Web.UnitTests/packages.config
@@ -17,14 +17,14 @@
   <package id="FluentAssertions" version="5.4.2" targetFramework="net462" />
   <package id="FluentValidation" version="8.0.0" targetFramework="net462" />
   <package id="Hashids.net" version="1.2.2" targetFramework="net452" />
-  <package id="MediatR" version="5.1.0" targetFramework="net462" />
+  <package id="MediatR" version="6.0.0" targetFramework="net462" />
   <package id="Microsoft.AspNet.Mvc" version="5.2.6" targetFramework="net462" />
   <package id="Microsoft.AspNet.Razor" version="3.2.6" targetFramework="net462" />
   <package id="Microsoft.AspNet.WebApi.Client" version="5.2.6" targetFramework="net462" />
   <package id="Microsoft.AspNet.WebApi.Core" version="5.2.6" targetFramework="net462" />
   <package id="Microsoft.AspNet.WebPages" version="3.2.6" targetFramework="net462" />
   <package id="Microsoft.AspNetCore.WebUtilities" version="2.1.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.DocumentDB.Core" version="2.1.3" targetFramework="net462" />
+  <package id="Microsoft.Azure.DocumentDB.Core" version="2.2.1" targetFramework="net462" />
   <package id="Microsoft.Azure.KeyVault.Core" version="1.0.0" targetFramework="net462" />
   <package id="Microsoft.Extensions.Primitives" version="2.1.1" targetFramework="net462" />
   <package id="Microsoft.IdentityModel.Clients.ActiveDirectory" version="3.19.8" targetFramework="net462" />
@@ -39,26 +39,27 @@
   <package id="NLog" version="4.3.8" targetFramework="net452" />
   <package id="NUnit" version="3.10.1" targetFramework="net452" />
   <package id="NUnit3TestAdapter" version="3.10.0" targetFramework="net462" />
-  <package id="SFA.DAS.Authorization" version="1.0.4" targetFramework="net462" />
-  <package id="SFA.DAS.AutoConfiguration" version="1.1.1" targetFramework="net462" />
+  <package id="SFA.DAS.Authorization" version="1.1.1" targetFramework="net462" />
+  <package id="SFA.DAS.AutoConfiguration" version="2.0.1" targetFramework="net462" />
   <package id="SFA.DAS.Commitments.Api.Client" version="2.1.716" targetFramework="net462" />
   <package id="SFA.DAS.Commitments.Api.Types" version="2.1.716" targetFramework="net462" />
   <package id="SFA.DAS.Common.Domain" version="1.4.43" targetFramework="net462" />
-  <package id="SFA.DAS.CosmosDb" version="1.0.1" targetFramework="net462" />
+  <package id="SFA.DAS.CosmosDb" version="1.0.47" targetFramework="net462" />
   <package id="SFA.DAS.CosmosDb.Testing" version="1.0.1" targetFramework="net462" />
   <package id="SFA.DAS.HashingService" version="1.0.0.42526" targetFramework="net452" />
   <package id="SFA.DAS.Http" version="1.2.4" targetFramework="net462" />
   <package id="SFA.DAS.Learners" version="1.0.0.39408" targetFramework="net452" />
   <package id="SFA.DAS.NLog.Logger" version="1.0.0.39208" targetFramework="net462" />
-  <package id="SFA.DAS.ProviderRelationships.Api.Client" version="1.1.1" targetFramework="net462" />
+  <package id="SFA.DAS.ProviderRelationships.Api.Client" version="2.0.1" targetFramework="net462" />
   <package id="SFA.DAS.ProviderRelationships.ReadStore" version="1.1.1" targetFramework="net462" />
-  <package id="SFA.DAS.ProviderRelationships.Types" version="1.1.1" targetFramework="net462" />
+  <package id="SFA.DAS.ProviderRelationships.Types" version="2.0.1" targetFramework="net462" />
   <package id="SFA.DAS.Testing" version="2.0.19" targetFramework="net462" />
   <package id="StructureMap" version="4.7.0" targetFramework="net462" />
   <package id="System.AppContext" version="4.3.0" targetFramework="net462" />
   <package id="System.Buffers" version="4.5.0" targetFramework="net462" />
   <package id="System.Collections" version="4.3.0" targetFramework="net462" />
   <package id="System.Collections.Concurrent" version="4.3.0" targetFramework="net462" />
+  <package id="System.Collections.Immutable" version="1.3.0" targetFramework="net462" />
   <package id="System.Collections.NonGeneric" version="4.0.1" targetFramework="net462" />
   <package id="System.Collections.Specialized" version="4.0.1" targetFramework="net462" />
   <package id="System.ComponentModel.Annotations" version="4.4.1" targetFramework="net462" />
@@ -82,7 +83,7 @@
   <package id="System.Linq.Expressions" version="4.3.0" targetFramework="net462" />
   <package id="System.Linq.Queryable" version="4.0.1" targetFramework="net462" />
   <package id="System.Memory" version="4.5.1" targetFramework="net462" />
-  <package id="System.Net.Http" version="4.3.3" targetFramework="net462" />
+  <package id="System.Net.Http" version="4.3.4" targetFramework="net462" />
   <package id="System.Net.NameResolution" version="4.0.0" targetFramework="net462" />
   <package id="System.Net.NetworkInformation" version="4.1.0" targetFramework="net462" />
   <package id="System.Net.Primitives" version="4.3.0" targetFramework="net462" />
