+++ src/SFA.DAS.ProviderApprenticeshipsService.Web/DependencyResolution/LinkGeneratorRegistry.cs
@@ -1,0 +1,16 @@
+﻿using FluentValidation;
+using SFA.DAS.ProviderApprenticeshipsService.Web.Models;
+using SFA.DAS.ProviderApprenticeshipsService.Web.Validation;
+using SFA.DAS.ProviderUrlHelper;
+using StructureMap;
+
+namespace SFA.DAS.ProviderApprenticeshipsService.Web.DependencyResolution
+{
+    public class LinkGeneratorRegistry : Registry
+    {
+        public LinkGeneratorRegistry()
+        {
+            For<ILinkGenerator>().Use<LinkGenerator>().Singleton();
+        }
+    }
+}
