+++ src/SFA.DAS.ProviderApprenticeshipsService.Web/Views/Account/Index.cshtml
@@ -44,7 +44,7 @@
                 {
                     <div class="box">
                         <h2 class="heading-medium">
-                            <a href="@UrlHelper.ProviderUrlHelperExtensions($"{Model.ProviderId}/reservations")">Manage reservations</a>
+                            <a href="@Url.ReservationsLink($"{Model.ProviderId}/reservations")">Manage reservations</a>
                         </h2>
                         <p>Manage reservations on behalf of an employer.</p>
                     </div>
