+++ src/SFA.DAS.ProviderApprenticeshipsService.Web/packages.config
@@ -19,7 +19,7 @@
   <package id="IdentityServer3" version="2.5.1" targetFramework="net45" />
   <package id="jQuery" version="1.10.2" targetFramework="net452" />
   <package id="jQuery.Validation" version="1.11.1" targetFramework="net452" />
-  <package id="MediatR" version="5.1.0" targetFramework="net462" />
+  <package id="MediatR" version="6.0.0" targetFramework="net462" />
   <package id="Microsoft.ApplicationInsights" version="2.5.1" targetFramework="net462" />
   <package id="Microsoft.ApplicationInsights.Agent.Intercept" version="2.4.0" targetFramework="net45" />
   <package id="Microsoft.ApplicationInsights.DependencyCollector" version="2.5.1" targetFramework="net462" />
@@ -38,7 +38,7 @@
   <package id="Microsoft.AspNet.WebApi.Core" version="5.2.6" targetFramework="net462" />
   <package id="Microsoft.AspNet.WebPages" version="3.2.6" targetFramework="net462" />
   <package id="Microsoft.AspNetCore.WebUtilities" version="2.1.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.DocumentDB.Core" version="2.1.3" targetFramework="net462" />
+  <package id="Microsoft.Azure.DocumentDB.Core" version="2.2.1" targetFramework="net462" />
   <package id="Microsoft.Azure.KeyVault.Core" version="1.0.0" targetFramework="net452" />
   <package id="Microsoft.CodeDom.Providers.DotNetCompilerPlatform" version="1.0.8" targetFramework="net462" />
   <package id="Microsoft.Data.Edm" version="5.8.0" targetFramework="net45" />
@@ -76,10 +76,10 @@
   <package id="Respond" version="1.2.0" targetFramework="net452" />
   <package id="SFA.DAS.Account.Api.Client" version="1.3.928" targetFramework="net462" />
   <package id="SFA.DAS.Account.Api.Types" version="1.3.928" targetFramework="net462" />
-  <package id="SFA.DAS.Authorization" version="1.0.4" targetFramework="net462" />
+  <package id="SFA.DAS.Authorization" version="1.1.1" targetFramework="net462" />
   <package id="SFA.DAS.Authorization.Mvc" version="1.0.4" targetFramework="net462" />
-  <package id="SFA.DAS.Authorization.ProviderPermissions" version="1.0.4" targetFramework="net462" />
-  <package id="SFA.DAS.AutoConfiguration" version="1.1.1" targetFramework="net462" />
+  <package id="SFA.DAS.Authorization.ProviderPermissions" version="1.1.1" targetFramework="net462" />
+  <package id="SFA.DAS.AutoConfiguration" version="2.0.1" targetFramework="net462" />
   <package id="SFA.DAS.Commitments.Api.Client" version="2.1.716" targetFramework="net462" />
   <package id="SFA.DAS.Commitments.Api.Types" version="2.1.716" targetFramework="net462" />
   <package id="SFA.DAS.Common.Domain" version="1.4.43" targetFramework="net462" />
@@ -87,7 +87,7 @@
   <package id="SFA.DAS.Configuration.AzureTableStorage" version="1.0.0.53229" targetFramework="net462" />
   <package id="SFA.DAS.Configuration.FileStorage" version="1.0.0.53229" targetFramework="net462" />
   <package id="SFA.DAS.CookieService" version="1.0.0.27423" targetFramework="net45" />
-  <package id="SFA.DAS.CosmosDb" version="1.0.1" targetFramework="net462" />
+  <package id="SFA.DAS.CosmosDb" version="1.0.47" targetFramework="net462" />
   <package id="SFA.DAS.CosmosDb.Testing" version="1.0.1" targetFramework="net462" />
   <package id="SFA.DAS.HashingService" version="1.0.0.42526" targetFramework="net45" />
   <package id="SFA.DAS.Http" version="1.2.4" targetFramework="net462" />
@@ -96,9 +96,9 @@
   <package id="SFA.DAS.NLog.Targets.Redis" version="1.0.0.34273" targetFramework="net45" />
   <package id="SFA.DAS.Notifications.Api.Client" version="2.0.0.35919" targetFramework="net45" />
   <package id="SFA.DAS.Notifications.Api.Types" version="2.0.0.35919" targetFramework="net45" />
-  <package id="SFA.DAS.ProviderRelationships.Api.Client" version="1.1.1" targetFramework="net462" />
+  <package id="SFA.DAS.ProviderRelationships.Api.Client" version="2.0.1" targetFramework="net462" />
   <package id="SFA.DAS.ProviderRelationships.ReadStore" version="1.1.1" targetFramework="net462" />
-  <package id="SFA.DAS.ProviderRelationships.Types" version="1.1.1" targetFramework="net462" />
+  <package id="SFA.DAS.ProviderRelationships.Types" version="2.0.1" targetFramework="net462" />
   <package id="SFA.DAS.Providers.Api.Client" version="0.10.29-prerelease" targetFramework="net45" />
   <package id="SFA.DAS.Sql.Client" version="1.0.0.32930" targetFramework="net45" />
   <package id="SFA.DAS.Web.Policy" version="1.1.0.44309" targetFramework="net45" />
@@ -112,6 +112,7 @@
   <package id="System.Buffers" version="4.5.0" targetFramework="net462" />
   <package id="System.Collections" version="4.3.0" targetFramework="net462" />
   <package id="System.Collections.Concurrent" version="4.3.0" targetFramework="net462" />
+  <package id="System.Collections.Immutable" version="1.3.0" targetFramework="net462" />
   <package id="System.Collections.NonGeneric" version="4.0.1" targetFramework="net462" />
   <package id="System.Collections.Specialized" version="4.0.1" targetFramework="net462" />
   <package id="System.ComponentModel.Annotations" version="4.4.1" targetFramework="net462" />
@@ -136,7 +137,7 @@
   <package id="System.Linq.Expressions" version="4.3.0" targetFramework="net462" />
   <package id="System.Linq.Queryable" version="4.0.1" targetFramework="net462" />
   <package id="System.Memory" version="4.5.1" targetFramework="net462" />
-  <package id="System.Net.Http" version="4.3.3" targetFramework="net462" />
+  <package id="System.Net.Http" version="4.3.4" targetFramework="net462" />
   <package id="System.Net.NameResolution" version="4.0.0" targetFramework="net462" />
   <package id="System.Net.NetworkInformation" version="4.1.0" targetFramework="net462" />
   <package id="System.Net.Primitives" version="4.3.0" targetFramework="net462" />
