+++ src/SFA.DAS.ProviderUrlHelperTests/SFA.DAS.ProviderUrlHelperTests.csproj
@@ -24,7 +24,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\SFA.DAS.ProviderUrlHelper\SFA.DAS.ProviderUrlHelper.csproj" />
-    <ProjectReference Include="..\SFA.DAS.UrlHelper\SFA.DAS.ProviderUrlHelper.csproj" />
   </ItemGroup>
 
 </Project>
